Add configuration option for allowed enchantments
Add a configuration option `allowedEnchantments` to allow players to choose which enchantments they want to use with the repair feature. The plugin will only use enchantments from this list for repair.

Modify the event handler to use the `containsKey` method of the `Map` class to check if an `ItemStack` has the MENDING enchantment, and the `contains` method of the `List` class to check if the MENDING enchantment is in the list of allowed enchantments.

Remove the `metrics` variable from the `onEnable` method, as it is never used.

Upgrade to the latest version of the bStats library to fix a compilation error.